+++ main/itf_sd_card_writer.h
@@ -7,6 +7,7 @@ int itf_addTestToSD(int testNum);
 void itf_writeSD_task(void * params);
 int itf_forceWriteBuffers(void);
 void itf_writeTestMessage(char *str);
+int itf_addImportantData(void);
 extern int itf_forceWriteBuffers_FLAG;
 
 #endif
